+++ src/display/json.rs
@@ -1,21 +1,26 @@
 use std::io::{self, Write};
 
-use chrono::{DateTime, NaiveDate, Utc};
+use chrono::NaiveDate;
 use serde::Serialize;
 
 use crate::{
-    display::renderer::Renderer,
-    models::{
-        dayfile::DayFile,
-        item::{Item, ItemPriority, ItemStatus},
+    display::{
+        json::dayfile_output::{DayFileOutput, Response},
+        renderer::Renderer,
     },
+    models::{dayfile::DayFile, item::Item},
 };
+
+mod dayfile_output;
+mod summary_output;
 
 pub struct JsonRenderer;
 
 impl Renderer for JsonRenderer {
     fn render_day(&self, df: &DayFile) -> Result<(), std::io::Error> {
-        Self::to_json(&DayFileOutput::from(df))
+        let output = &DayFileOutput::from(df);
+        let response = Response::<&DayFileOutput>::new("ls", output);
+        Self::to_json(&response)
     }
 
     fn render_summary(&self, _index: Option<usize>, item: &Item) -> Result<(), std::io::Error> {
@@ -63,84 +68,5 @@ impl JsonRenderer {
         writeln!(out)?;
 
         Ok(())
-    }
-}
-
-// Models
-const SCHEMA_VERSION: u8 = 1;
-
-#[derive(Serialize, Debug)]
-struct DayOutput {
-    date: NaiveDate,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    path: Option<String>,
-}
-
-#[derive(Serialize, Debug)]
-struct ItemOutput {
-    id: String,
-    text: String,
-    created_at: DateTime<Utc>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    done_at: Option<DateTime<Utc>>,
-    priority: ItemPriority,
-    tags: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    due: Option<DateTime<Utc>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    notes: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    migrated_from_date: Option<NaiveDate>,
-    status: ItemStatus,
-}
-
-#[derive(Serialize, Debug)]
-struct SummaryOutput {
-    total: usize,
-    open: usize,
-    done: usize,
-}
-
-#[derive(Serialize, Debug)]
-struct DayFileOutput {
-    schema_version: u8,
-    // command: &'static str,
-    day: DayOutput,
-    summary: SummaryOutput,
-    items: Vec<ItemOutput>,
-}
-
-impl From<&Item> for ItemOutput {
-    fn from(value: &Item) -> Self {
-        Self {
-            id: value.id.clone(),
-            text: value.text.clone(),
-            created_at: value.created_at,
-            done_at: value.done_at,
-            priority: value.priority,
-            tags: value.tags.clone(),
-            due: value.due,
-            notes: value.notes.clone(),
-            migrated_from_date: value.migrated_from,
-            status: value.status(),
-        }
-    }
-}
-
-impl From<&DayFile> for DayFileOutput {
-    fn from(value: &DayFile) -> Self {
-        let total = value.items.len();
-        let done = value.items.iter().filter(|i| i.done_at.is_some()).count();
-        let open = total - done;
-
-        Self {
-            schema_version: SCHEMA_VERSION,
-            day: DayOutput {
-                date: value.date,
-                path: None,
-            },
-            summary: SummaryOutput { total, open, done },
-            items: value.items.iter().map(ItemOutput::from).collect(),
-        }
     }
 }
